pyproject.toml
[project]
name = "vyw-backend-new"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.13.12"
dependencies = [
    "aiosmtplib>=5.1.0",
    "alembic>=1.18.4",
    "fastapi[standard]>=0.135.3",
    "flash-attn==2.8.3",
    "greenlet>=3.3.2",
    "huggingface-hub>=0.36.2",
    "pillow>=12.1.1",
    "psycopg[binary]>=3.3.3",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "qwen-tts>=0.1.1",
    "sox>=1.5.0",
    "sqlalchemy>=2.0.49",
    "torch>=2.11.0",
    "torchaudio>=2.11.0",
]

[tool.uv.sources]
flash-attn = { url = "https://github.com/mjun0812/flash-attention-prebuild-wheels/releases/download/v0.9.6/flash_attn-2.8.3+cu128torch2.11-cp313-cp313-win_amd64.whl" }
torch = { index = "pytorch" }
torchaudio = { index = "pytorch" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu128"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]
